requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asgi-correlation-id==4.3.1
attrs==23.2.0
bcrypt==4.1.1
bytecode==0.15.1
cattrs==23.2.3
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
ddsketch==3.0.1
ddtrace==2.8.1
Deprecated==1.2.14
ecdsa==0.18.0
envier==0.5.1
fastapi==0.104.1
h11==0.14.0
httptools==0.6.1
idna==3.6
importlib-metadata==7.0.0
kink==0.7.0
opentelemetry-api==1.24.0
passlib==1.7.4
protobuf==5.26.1
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pyparsing==3.1.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
setuptools==69.5.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.5.0
starlette==0.27.0
structlog==24.1.0
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
wrapt==1.16.0
xmltodict==0.13.0
zipp==3.18.1
